fix: align SDK types with OpenAPI spec across 10 files
- Add 'type' discriminator with DiscriminatorValue to Event union (47 variants)
- Add missing input/raw fields to ToolStatePending
- Fix SessionSummary required/optional field tags
- Type StructuredOutputError, ContextOverflowError Data as proper structs
- Add missing Metadata field to APIError Data
- Add BadRequestError and NotFoundError to shared types + aliases
- Add typed enums for SnapshotFileDiffStatus, toast Variant, workspace Status
- Fix GlobalSession.Permission from interface{} to []PermissionRule
- Fix number type mismatches (int64 -> float64) in provider, pty, experimentalsession

+++ pty.go
@@ -206,8 +206,8 @@ func (r PtyUpdateParams) URLQuery() (v url.Values) {
 }
 
 type PtyUpdateSize struct {
-	Rows param.Field[int64] `json:"rows,required"`
-	Cols param.Field[int64] `json:"cols,required"`
+	Rows param.Field[float64] `json:"rows,required"`
+	Cols param.Field[float64] `json:"cols,required"`
 }
 
 type PtyRemoveParams struct {
